Move random loadout to a component, clean it up
