+++ src/index.js
@@ -34,7 +34,6 @@ require('./scripts/services/dimItemInfoService.factory');
 
 require('./scripts/loadout/dimLoadout.directive');
 require('./scripts/loadout/dimLoadoutPopup.directive');
-require('./scripts/loadout/random/dimRandom.controller');
 require('./scripts/shell/dimAngularFilters.filter');
 require('./scripts/shell/dimSearchFilter.directive');
 require('./scripts/shell/dimClickAnywhereButHere.directive');
