+++ src/scripts/loadout/random/random-loadout.html
@@ -1,0 +1,3 @@
+<div class="random-loadout" ng-if="::$ctrl.showRandomLoadout">
+  <a class="loadout random" href="#" ng-disabled="$ctrl.disableRandomLoadout" ng-click="$ctrl.applyRandomLoadout($event)">&Pi;</a>
+</div>
