+++ src/scripts/loadout/random/random-loadout.scss
@@ -1,0 +1,10 @@
+.random {
+  position: fixed;
+  backface-visibility: hidden;
+  bottom: 0;
+  right: 0;
+  padding: .5rem;
+  color: #222;
+  text-decoration: none;
+  font-weight: normal;
+}
