+++ src/scripts/store/inventory.html
@@ -6,6 +6,4 @@
 <div id="drag-help" class="drag-help-hidden" translate="Help.Drag"></div>
 <dim-farming></dim-farming>
 <dim-clear-new-items></dim-clear-new-items>
-<div class="random-loadout" ng-controller="dimRandomCtrl as random">
-  <a class="loadout random" href="#" ng-disabled="random.disableRandomLoadout" ng-if="::random.showRandomLoadout" ng-click="random.applyRandomLoadout($event)">&Pi;</a>
-</div>
+<random-loadout></random-loadout>
