+++ src/scripts/store/inventory.module.js
@@ -1,10 +1,12 @@
 import angular from 'angular';
 
 import inventoryComponent from './inventory.component';
+import { RandomLoadoutComponent } from '../loadout/random/random-loadout.component';
 
 export default angular
   .module('inventoryModule', [])
   .component('inventory', inventoryComponent)
+  .component('randomLoadout', RandomLoadoutComponent)
   .config(($stateProvider) => {
     'ngInject';
