+++ src/scss/_style.scss
@@ -1629,19 +1629,6 @@ h2, h3, h4 {
   }
 }
 
-.loadout {
-  &.random {
-    position: fixed;
-    backface-visibility: hidden;
-    bottom: 0;
-    right: 0;
-    padding: .5rem;
-    color: #222;
-    text-decoration: none;
-    font-weight: normal;
-  }
-}
-
 .current.character .character-box:before {
   content: '';
   position: absolute;
